modemfwd: Allow loading FW from rootfs even if DLC exists
This change allows a variant to reserve DLC space for future use. This
is required for boards like herobrine, where rootfs is the preferred
location of the modem fw, but we still want to reserve space for DLCs.

BUG=b:245410656
TEST=DLC is installed on hoglin, but FW is loaded from rootfs
TEST=FEATURES=test emerge-herobrine modemfwd

Cq-Depend: chromium:3952973, chromium:3950052, chrome-internal:5031328, chrome-internal:5031331, chrome-internal:5030973
Change-Id: Ia5653e57a6ef49791dad8a9d6e38dc7b05322cd1
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3938456
Reviewed-by: Andrew Lassalle <andrewlassalle@chromium.org>
Tested-by: Pavan Holla <pholla@google.com>
Commit-Queue: Pavan Holla <pholla@google.com>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>