resourced: Notify power supply status change to resourced
Currently resourced only checks the AC/DC mode when an activity changes.
This means that if the power state changes while an activity is active,
it won't update the settings.  With this patch, udevd would notify
resourced to update the power preferences.

The power supply change notification udev rule is not for permission.
Rename 99-resourced-permissions.rules to 99-resourced.rules as it
contains general udev rules.

BUG=b:246770820
TEST=tast run $DUT_IP platform.Resourced.baseline

Cq-Depend: chromium:3943268
Change-Id: I410f2a47771e8c5c2015bc2c0b3b809aaeed3cd3
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3935740
Reviewed-by: Raul Rangel <rrangel@chromium.org>
Tested-by: Vovo Yang <vovoy@chromium.org>
Commit-Queue: Vovo Yang <vovoy@chromium.org>