vm_tools: cros_im: Fix cursor location for X11 apps
The cursor location we report is supposed to be relative to the surface,
but is calculated with gdk_window_get_origin, which gives root window
coordinates. This seems to work under Wayland (as GDK does not actually
know where on the screen it is), but tends to be incorrect under X11.

This CL subtracts the offset of the top-level GDK window (appears to be
zero under Wayland) to fix this for X11 apps (crrev.com/c/3934678 is
also needed to make sommelier correctly scale the offset we provide).

BUG=b:232048354
TEST=Manual. Apply crrev.com/c/3934678, launch wayland and x11 apps with
various scales (explicitly passing --scale to sommelier) and verify that
IME UI (e.g. the emoji keyboard) is correctly positioned next to the
cursor.

Change-Id: I747d0a178aa1be51612fb3d59a10c3d901e65e14
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3941327
Reviewed-by: kokoro <noreply+kokoro@google.com>
Tested-by: kokoro <noreply+kokoro@google.com>
Reviewed-by: Sophia Lin <sophialin@google.com>
Commit-Queue: Timothy Loh <timloh@chromium.org>