vm_tools: cros_im: Update surrounding text when cursor location changes
Chrome expects that set_surrounding_text is called whenever there are
changes to the surrounding text (e.g. the user types something or moves
the cursor). This CL updates cros_im's GTK IM module so that whenever we
get set_cursor_location called, we update the surrounding text. Although
the name doesn't suggest that it'll be triggered for all updates that
may affect surrounding text, in practice it appears to be correctly
triggered (and there isn't anything else for us to observe anyway).

BUG=b:232048095,b:250810065
TEST=ninja && ../test/run_tests.py

Change-Id: I4583db326abd6dbe1b4762ae75b1f04c4f91fbef
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3941331
Reviewed-by: kokoro <noreply+kokoro@google.com>
Tested-by: kokoro <noreply+kokoro@google.com>
Reviewed-by: Sophia Lin <sophialin@google.com>
Commit-Queue: Timothy Loh <timloh@chromium.org>