mojo_service_manager: Add madvise to seccomp policies
To enable scudo to work with mojo_service_manager, it
must be able to call the madvise syscall directly with
MADV_DONTNEED. At present this is forbidden, but we can allow
it here.

It may just be better to use madvise: 1, as that's done in
a number of syscall filters, but it's occasionally been given
higher restrictions, and scudo only calls it with this argument.

BUG=b:248085566
TEST=Built and ran volteer image with scudo
TEST=CQ

Change-Id: I4eadead56e8f8599b24f952d5d2ded3884a60c23
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3944054
Tested-by: Jordan Abrahams-Whitehead <ajordanr@google.com>
Reviewed-by: Chung-sheng Wu <chungsheng@google.com>
Reviewed-by: Zi Lin <lziest@google.com>
Reviewed-by: Ryo Hashimoto <hashimoto@chromium.org>
Commit-Queue: Jordan Abrahams-Whitehead <ajordanr@google.com>