faced: Import camera handling code from internal faceauth repo
File paths of `#include` lines and header guards have been updated, but
otherwise the code is imported without change.

The code does not compile at this stage, nor is it intended to. This
change acts as a baseline for future CLs which update the code to
simplify review and history.

BUG=b:237233809
TEST=None

Change-Id: I10540e79ae7915f73f01010735574103d94cbac4
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3945488
Reviewed-by: Victor Wong <vwon@google.com>
Tested-by: David Greenaway <dgreenaway@google.com>
Commit-Queue: David Greenaway <dgreenaway@google.com>