vm_tools: cros_im: Don't send extremely long surrounding text
Due to the Wayland protocol's message size limit, we need to limit the
size of text sent over text_input::set_surrounding_text. This CL just
makes us not send the message if it exceeds 4000 bytes (actual limit
appears to be ~4075 bytes), although we should in the future fix this to
return an appropriate substring.

BUG=b:232048095
TEST=Manual: Insert >4KB of text on a single line in gedit and check
there is no crash.

Change-Id: I68038ee9c60ce50766ef00282c9c7a7ba217f349
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3944271
Reviewed-by: kokoro <noreply+kokoro@google.com>
Commit-Queue: Timothy Loh <timloh@chromium.org>
Tested-by: Timothy Loh <timloh@chromium.org>
Reviewed-by: Sophia Lin <sophialin@google.com>