shill: net: Create IPAddress from string containing prefix
This patch added the function to create the IP address object from the
string that has an IP address and prefix. If the string neither has the
appropriate format for IPv4 nor IPv6, this function will create an IP
address object whose family is unknown for an error.

BUG=b:251686186
TEST=unit: FEATURES="test" emerge-$BOARD shill-net

Change-Id: Iafb2d69efb0edf8b5e42573e8c470ddeb8f543c0
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3944307
Reviewed-by: Jie Jiang <jiejiang@chromium.org>
Reviewed-by: Taoyu Li <taoyl@chromium.org>
Commit-Queue: Arisa Tsuji <arisai@google.com>
Tested-by: Arisa Tsuji <arisai@google.com>