typecd: Report power source location metrics
Report power source location metrics on partner add. It is only reported
if the port's power role is sink, or the partner is a power source.

PortManager keeps track of the port number that was previously used as a
sink for charging the system.

BUG=b:250978699
TEST=FEATURES=test emerge-brya chromeos-base/typecd

Cq-Depend: chromium:3947382
Change-Id: I16c7a7b5ef7ba48e93a8f583d52419d111283fab
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3947383
Reviewed-by: Benson Leung <bleung@google.com>
Commit-Queue: Won Chung <wonchung@google.com>
Tested-by: Won Chung <wonchung@google.com>