cryptohome: Handle ephemeral user existence correctly in AuthSession
The existing code doesn't handle "user exists" correctly for ephemeral
AuthSessions. When you start a session for a new user and create the
ephemeral mount, the session will then report user_exists == true. But
on subsequent AuthSessions for the same users, it will go back to
reporting false because on construction an AuthSession tests for
existence by looking for the persistent homedir.

This fix works by generalizing the check to either "there is a
persistent homedir, or there is an active mounted homedir".

BUG=b:251508938
TEST=FEATURES=test emerge-$BOARD cryptohome
TEST=manually tested on machine

Change-Id: I311a434a07cf03c53d4abd23558ae90f8b56c8ad
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3946225
Reviewed-by: Hardik Goyal <hardikgoyal@chromium.org>
Tested-by: John Admanski <jadmanski@chromium.org>
Commit-Queue: John Admanski <jadmanski@chromium.org>
Reviewed-by: Betul Soysal <betuls@google.com>