crosh: Update Downloads references to not use bind mount
The ~/Downloads directory is bind mounted to ~/MyFiles/Downloads, this
folder is being migrated in M109 and will no longer be a valid
reference. Update to point to ~/MyFiles/Downloads instead.

NOTE: I've bypassed hooks on this as there are a number of lines that
are throwing errors in shell and rust that I am not familiar with
(unrelated to my change)

BUG=b:253122024
TEST=CQ

Change-Id: Ifef1d7c4fe0a533dea4b7b851ce63cf2bba014e6
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3947923
Commit-Queue: Ben Reich <benreich@chromium.org>
Tested-by: Ben Reich <benreich@chromium.org>
Reviewed-by: Allen Webb <allenwebb@google.com>