camera: pick face detection buffer per request
Auto-framing configures extra YUV streams and dynamically selects the
active YUV stream based on the feature activation state. This means we
can no longer assume we have a fixed preview stream, hence we need to
modify FaceDetection SM to pick the YUV buffer to process on a
per-request basis.

While we're at it, also fix the coordinates transform for the bounding
box and landmark coordinates in the feature metadata.

BUG=None
TEST=Enable frame annotator and verify the boxes and landmarks drawn on
are correct when auto-framing is turned on and the frame is cropped.

Change-Id: I955a074d446f0ad52f17517ac25bb7abb5aadc66
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3954403
Tested-by: Ricky Liang <jcliang@chromium.org>
Reviewed-by: Ren-Pei Zeng <kamesan@chromium.org>
Commit-Queue: Ricky Liang <jcliang@chromium.org>