shill: Check state in Network::OnStaticIPConfigChanged()
Check state instead of the existence of ipconfig which should be more
correct.

This patch will change the behavior of the Network class. After this
patch, the Network class will be connected by StaticIPConfig even if
DHCP is not started. This is expected and will not actually affect any
current users of Network: only Cellular and VPN may not set DHCP and
Network will be set up in Network::Start() directly in both cases.

Also improve the unit test to make sure Network will not be connected by
a partial StaticIPConfig.

BUG=b:235330956
BUG=b:232177767
TEST=FEATURES="test" emerge-${BOARD} shill

Change-Id: Ic7d70f2a68a70438e15c7db21c0bbe8eadbf62ae
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3949070
Reviewed-by: Taoyu Li <taoyl@chromium.org>
Commit-Queue: Jie Jiang <jiejiang@chromium.org>
Tested-by: Jie Jiang <jiejiang@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>