cryptohome: Replace TakeCredentials with set_key_data
Note: this is a reland of chromium:3949959

With this function no longer grabbing verifiers, we can change the flow
to instead push the "grab keydata from auth, set it in used" directly
into the calling code. This simplifies things a little, and the code in
question can go away entirely once we can prune the direct use of
KeyData after the old mount APIs are gone.

This simplifies the verifier flow a little more, and it also lets us
break the soft dependency of UserSession on AuthSession, a dependency
that mostly flows in the other direction.

BUG=b:247812443
TEST=FEATURES=test emerge-$BOARD cryptohome

Change-Id: Ib2fc7bfa078e036228b8627d09f3ad97c4fa9135
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3960555
Tested-by: John Admanski <jadmanski@chromium.org>
Reviewed-by: Betul Soysal <betuls@google.com>
Commit-Queue: John Admanski <jadmanski@chromium.org>