shill: check existing address as the first step in UpdateFromIPConfig
In Connection::UpdateConfig, we currently check if any address other
than the shill-managed one was configured on the netdevice, and flush
all addresses and routes if detected.

This operation should be done as the first step in UpdateFromIPConfig.
Otherwise the result of previous steps (currently, SetupExcludedRoutes()
and FixGatewayReachability()) gets lost during the flush.

BUG=chromium:1364133
TEST=unit

Change-Id: I0554816b29179eff150fce0c017273fdf133c153
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3950270
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Tested-by: Taoyu Li <taoyl@chromium.org>
Reviewed-by: Jie Jiang <jiejiang@chromium.org>
Commit-Queue: Taoyu Li <taoyl@chromium.org>