resourced: Trim the trailing newline before parsing to u32
In set_global_ondemand_governor_value(), it fails to parse the file
content to u32 because there is trailing newline in the end of the
file. Trim the trailing newline to parse the file content to u32.

Simulate the trailing newline behavior of sysfs in unit tests to catch
this error.

BUG=b:253328282
TEST=tast run $DUT_IP platform.Resourced.baseline

Change-Id: Ia7ecfb70e25b6ccdbca1a9d266f74a7e8346106b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3951630
Reviewed-by: Charles William Dick <cwd@google.com>
Commit-Queue: Vovo Yang <vovoy@chromium.org>
Tested-by: Vovo Yang <vovoy@chromium.org>
Reviewed-by: Kameron Lutes <kalutes@chromium.org>