shill: Set pid in ProcessManagerTest
Currently, in ProcessManagerTest.StartProcessWithMinijailWithStdout, we
don't update the pid pointer when handling RunEnvPipesAndDestroy(), and
thus the local variable |pid| in
ProcessManager::StartProcessInMinijailWithPipesInternal() is
uninitialized. Sometimes it will have value -1 and we treat this value
as a signal of failure so the unit test will fail.

This patch fixes this issue.

BUG=b:253349908
TEST=FEATURES="test" emerge-${BOARD} shill

Change-Id: I59cdd3243b3559d7e003d0a436f0b69df1e7f8cc
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3951750
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Reviewed-by: Taoyu Li <taoyl@chromium.org>
Commit-Queue: Jie Jiang <jiejiang@chromium.org>
Tested-by: Jie Jiang <jiejiang@chromium.org>