camera: usb: Recover USB camera firmware version after running camera_dfu_test
The V4L2 and DFU tests are run in the same suite in AVL qualification testing, and the DFU test can change the FW version and causes unexpected results.
Recover to the original FW version after the test to ensure the consistency before and after the test.

BUG=b:249211799
TEST=camera_dfu_test \
    --app_mode_id=04f2:b741 \
    --dfu_mode_id=0bda:5800 \
    --fw1_path=/etc/camera/realtek/rts5846/RTS5846R_FW_Cache_AB_new_Suffix_V2222.bin \
    --fw1_version=2222 \
    --fw2_path=/etc/camera/realtek/rts5846/RTS5846R_FW_Cache_AB_new_Suffix_V1111.bin \
    --fw2_version=1111 \
    --quirks=detach-for-attach

Change-Id: I81ad3aefb35b4a2757014f58eef7e5b391205fcf
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3949097
Reviewed-by: Ren-Pei Zeng <kamesan@chromium.org>
Tested-by: XING-GU CHEN <xinggu@google.com>
Commit-Queue: XING-GU CHEN <xinggu@google.com>
Reviewed-by: Shik Chen <shik@chromium.org>