dlcservice: Avoid big writes when resuming from hibernate
When resuming from hibernate, we are attempting to log the user in
without incurring too many writes. This is because LVM is operating in
mostly read-only mode, and writes to the stateful partition are being
routed to a dm-snapshot stored in the hibervol. This snapshot is fairly
limited in size, so we need dlcservice to avoid big file system writes
to stateful, as well as avoid attempts to create new logical volumes.

When resuming from hibernate, fail to create new LVs for DLCs or
create+expand files that don't already exist.

BUG=b:237090545
TEST=delete dlc, fake in progress resume, observe dlc install fail

Change-Id: I5690ba755bcc7db2a622fe535e8f4dd02a9529fb
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3953398
Commit-Queue: Evan Green <evgreen@chromium.org>
Tested-by: Evan Green <evgreen@chromium.org>
Reviewed-by: Jae Hoon Kim <kimjae@chromium.org>