biod: Impose timeout on flashrom with a 2m duration
Add timeout for the CrosFpDeviceUpdate with a duration of 2 minutes.
Create a function GetAppOutputAndErrorWithTimeout, which returns status
& error and exists with a false status if the function times out. The
timeout command is prepended with a verbose flag in order to diagnose
any signal sent upon timeout. A preserve-status flag is added in order
to exit with the same status as the base command status. A kill-after
flag is added with a given duration to kill the signal if the command
is still running. Create unit tests for the function.

BUG=b:130026657
TEST=tast run <DUT> firmware.FpUpdater
TEST=cros_workon_make --board=hatch --test biod

Change-Id: I1a5b06076f916c77ab596f0b784b934ef8aa60cd
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3956559
Commit-Queue: Firas Sammoura <fsammoura@google.com>
Reviewed-by: Craig Hesling <hesling@chromium.org>
Tested-by: Firas Sammoura <fsammoura@google.com>