vtpm: refactor get_capability_command.cc
To support more capability types, extracting the TPM handles part into a
subfunction.

Also, we wanna know which unsupported TPM_CAP is queried, so this CL also
adds more logs to tell the TPM_CAP if it is not supported or just
unexpected.

This is a purely refactoring function expect for the increased
verbosity mentioned above.

BUG=b:251638804
TEST=FEATURES=test emerge-brya vtpm

Change-Id: Ide9adc2f45872bd5b34afa9eac09d5eca18e53e6
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3958323
Reviewed-by: Yi Chou <yich@google.com>
Commit-Queue: Leo Lai <cylai@google.com>
Tested-by: Leo Lai <cylai@google.com>