faced: Introduce CameraManager class
CameraManager provides a simplified interface allowing callers to start
and stop camera streams. In particular, it allows creation of
a `StreamReader` object that will stream camera frames, and moves
responsibility for choice of format, resolution, frame rate, etc to the
creator of the CameraManager, instead of the party who opens the camera
stream.

BUG=b:237233809
TEST=`cros_workon_make --board=$BOARD --test chromeos-base/faced`

Change-Id: If9b3a972540d93aba83175e6a8bfe8566c671a30
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3966094
Commit-Queue: David Greenaway <dgreenaway@google.com>
Reviewed-by: Ben Kersten <bkersten@google.com>
Reviewed-by: Victor Wong <vwon@google.com>
Tested-by: David Greenaway <dgreenaway@google.com>