cryptohome: Get rid of CredentialVerifier::Set
As part of trying to unify credential verifiers with the auth block
verifier path, I need to refactor the verification to use AuthInput
instead of a blob secret. However, this is a lot cleaner if we only need
to do this with Verify and not Set.

I think it doesn't make a lot of sense for Set to be a part of the
generic interface, as verifier construction is already subclass-specific
and so you don't need a fully generic Set. I also don't think we really
need the verifiers to be mutable (we don't take advantage of this today,
ScryptVerifier is only ever Set once, immediately after constructing it.

I also added in a mock verifier implementation, so that unit tests which
just want a generic verifier (and which aren't even using Verify) don't
need to fake out a password-based scrypt one.

Plus on top of that I also made a few little style tweaks and cleanups
while I was touching the files. Like making the default and deleted
functions more style guide compliant, or tweaking the constructor
parameters to use a value-move passing style. And splitting the scrypt
verifier specific tests out into a separate scrypt_verifier_unittest.

BUG=b:247812443
TEST=FEATURES=test emerge-$BOARD cryptohome

Change-Id: Ic2b26c2abc6b389ec9b9a7568552b4aa0c653366
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3957456
Tested-by: John Admanski <jadmanski@chromium.org>
Reviewed-by: Thomas Cedeno <thomascedeno@google.com>
Reviewed-by: Betul Soysal <betuls@google.com>
Commit-Queue: John Admanski <jadmanski@chromium.org>