shill: Add ApnType to APN info
Replace is_attach_apn with apn_types to allow an APN to have multiple
types. This is a requirement to support DUN, FOTA and other APN types.
Since is_attach_apn means that an APN is also an InitialAttach APN, the
IA type is added to accomplish the same result.

BUG=b:253668588
TEST=tast run redrix  cellular.*Apn*

Change-Id: Ic296cbd5b2b7067889fe03463ef255cf1f0e27e7
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3956643
Reviewed-by: Aleksander Morgado <aleksandermj@google.com>
Tested-by: Andrew Lassalle <andrewlassalle@chromium.org>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>
Commit-Queue: Andrew Lassalle <andrewlassalle@chromium.org>