vm_tools: pass suggested username from vmc to vm
This modifies the StartVm and ConfigureContainerGuest interfaces to also
provide a `vm_username` field that is derived from the user's email
address. This matches the value passed to configure crostini containers.

This only applies to VMs that support specifying a username, such as
Bruschetta.

BUG=b:249381463
TEST=deployed to brya, confirmed crostini still works, confirmed the
username correctly shows up in Bruschetta

Change-Id: Ie1c7e40a604249de0870d34cb88e53666c133e40
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3958056
Reviewed-by: Fergus Dall <sidereal@google.com>
Commit-Queue: Mike Gerow <gerow@google.com>
Tested-by: Mike Gerow <gerow@google.com>