faced: Simplify the CameraClient device query API
In preparation of adding a FakeCameraClient, we simplify a CameraClient
query API. In particular, instead of providing functions that query
particular resolutions, provide an API that returns the set of available
devices and formats.

We additionally simplify the implementation of how the default
CameraClient implementation queries this information, relying on the
fact that the `GetCameraInfo` callback is synchronously called once for
device on the system when it is first called. This helps us to avoid
having to deal with asynchronous callbacks after we've finished querying
the device.

BUG=b:237233809
TEST=`cros_workon_make --board=$BOARD --test chromeos-base/faced`

Cq-Depend: chrome-internal:5034390
Change-Id: Ie7c13636d6371f9aff2753cee6b71d1b58538399
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3955738
Commit-Queue: David Greenaway <dgreenaway@google.com>
Reviewed-by: Victor Wong <vwon@google.com>
Reviewed-by: Ben Kersten <bkersten@google.com>
Tested-by: David Greenaway <dgreenaway@google.com>