avtest_label_detect: support 8K video deco
This CL adds support for detecting 8K accelerated video decoding
detection to said binary. It basically piggybacks on top of what's
existing for 4K, cleaning up a bit the code (e.g. removing superfluous
comments etc).

The whole folder seems to be pretty redundant, and due to be migrated
to C++, so I tried to only cut as much as needed.

This CL should land at the same time as crrev.com/c/3975898, which
updates the per-SoC capabilities files, so that video.Capability
(which compares the output of avtest_label_detect with the static
capabilities) doesn't fail.

BUG=b:250099630
TEST=On Brya

Cq-Depend: chromium:3975898
Change-Id: I0c15b495014a20a66109248d9d0cc2b55f99df2d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3960556
Reviewed-by: Hirokazu Honda <hiroh@chromium.org>
Tested-by: Miguel Casas-Sanchez <mcasas@chromium.org>
Commit-Queue: Miguel Casas-Sanchez <mcasas@chromium.org>