secanomalyd: adds the text file reader library
The text file reader class introduces file reading capabilities to
secanomalyd. The main use case for this is to be able to read from
audit log files, which is useful for detecting certain anomalies.
The class operates by keeping a file open and reading new lines from the
file when they are written to it. It remembers its position in the file
and accounts for file rotations, which is common in log files.

BUG=b:254286602
TEST=TexfFileReader tests pass. Secanomalyd tests pass.

Change-Id: I6260042cd63c7bcf1ba215a80640e131e4134fd6
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3964888
Tested-by: Shervin Oloumi <enlightened@chromium.org>
Reviewed-by: Jorge Lucangeli Obes <jorgelo@chromium.org>
Reviewed-by: Jeff Xu <jeffxu@google.com>
Commit-Queue: Shervin Oloumi <enlightened@chromium.org>