init: invoke storage specific logical erasure
Instead of calling BLKZEROOUT directly in WipeBlockDevice, we should
call the device specific logical erasure function if supported.
For UFS, it supports BLKDISCARD. For eMMC, it supports BLKSECDISCARD.
The rest of the storage types will still call BLKZEROOUT.
Physical erasure (purge for UFS and sanitize for eMMC) will be added in
the future.

BUG=b:229168577
TEST=echo "keepimg" > /mnt/stateful_partition/factory_install_reset

Change-Id: I7e9724b441ff23135f9e0a1d328b95d71e049ae1
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3984854
Tested-by: Phoebe Wang <phoebewang@chromium.org>
Commit-Queue: Phoebe Wang <phoebewang@chromium.org>
Reviewed-by: Gwendal Grignou <gwendal@chromium.org>
Reviewed-by: Daniil Lunev <dlunev@chromium.org>