vm_tools: cros_im: Remove BackendTest::CreateTextInputOptions enum
The BackendTest::CreateTextInputOptions enum currently allows tests to
either ignore several requests which are often called and tend to not be
relevant to most tests, or ignore no requests. The latter option is not
great as it means new requests added to the default path need to be
added to tests using that option.

This test removes this enum and makes all ExpectCreateTextInput calls
ignore the same set of requests. Tests can use Unignore() to apply
expectations for these requests if needed.

BUG=b:252730277
TEST=../test/run_tests.py

Change-Id: Ic465dd19f13da3e8a4e247ee8d285826ab985952
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3983515
Reviewed-by: Sophia Lin <sophialin@google.com>
Commit-Queue: Timothy Loh <timloh@chromium.org>
Reviewed-by: kokoro <noreply+kokoro@google.com>
Tested-by: kokoro <noreply+kokoro@google.com>