cryptohome: Have credential Verify use AuthInput.
This makes the verifier API more general, so that we could use it for
non-password credentials. This brings it pretty close to the
AuthBlockUtility verify API in terms of parity, with the one other
missing feature being a lack of async support which is still a bit of a
challenge so long as we have the old Mount API codepaths.

BUG=b:247812443
TEST=FEATURES=test emerge-$BOARD cryptohome

Change-Id: Id248ba3866f55bf99495a70feaf0e68f4ea8f5ef
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3964375
Commit-Queue: John Admanski <jadmanski@chromium.org>
Reviewed-by: Maksim Ivanov <emaxx@chromium.org>
Tested-by: John Admanski <jadmanski@chromium.org>
Reviewed-by: Thomas Cedeno <thomascedeno@google.com>