missive: ReportQueueProvider split
Requests queue and weak ptr factory expect that they are destruct on the
seq task runner. To facilitate it, split portion of the provider off and
do the destruction on thread.
To make sure future changes do not break it, DCHECK(sequence_checker) is
added to the destructor.

BUG=b:254310475
TEST=FEATURES="test" emerge-$BOARD missive

Change-Id: I09532f312e8e2ddff419684ce0693dd5d4c0cd76
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3964376
Auto-Submit: Leonid Baraz <lbaraz@chromium.org>
Reviewed-by: Vignesh Shenvi <vshenvi@google.com>
Tested-by: Leonid Baraz <lbaraz@chromium.org>
Commit-Queue: Leonid Baraz <lbaraz@chromium.org>