hiberman: Add resume_in_progress token file
In resume-init, add a tmpfs file at /run/hibernate/resume_in_progress
that indicates the system is undergoing a resume from hibernate. This
can be used by other services that need to check to see if a resume is
underway, without needing to add an upstart or dbus dependency on
hiberman. When resume is aborted, the cookie file is deleted. As a tmpfs
file, it won't stick around across an unexpected reboot or power cycle.

BUG=b:237090293
TEST=hibernate, resume, and abort, observe cookie file

Change-Id: Ia5c30e9acf9ca5f365595d1420dd387025a2d505
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3969697
Reviewed-by: Alexis Savery <asavery@chromium.org>
Tested-by: Evan Green <evgreen@chromium.org>
Commit-Queue: Evan Green <evgreen@chromium.org>