shill: tethering: isolate load/unload from Profiles
This patch changes the logic in TetheringManager when loading or
unloading a configuration from a user profile storage such that
TetheringManager does not need to rely on global shill state. Instead
the top-level manager calls Load and Unload appropriately based on
the type of profile being pushed or popped.

The main benefit is to allow TetheringManager to ignore entirely the
state of shill profiles when managing the config, separating concerns
better.

This allows for instance to test configuration load and save atomically
in unit tests. Manager unit tests checking interactions with
TetheringManager are also added.

A follow-up patch will do the same transformation for SetEnabled and
SetAndPersistConfiguration.

This patch also introduces various helpers for simpliflying and
streamlinming TetheringManager unit tests and fixes a flaky test.

BUG=b:248028325
BUG=b:253208938
TEST=FEATURES=test emerge-$BOARD shill

Change-Id: Iee49466fd1a87ea4b9ff4d22ec7adcb52cb48c5c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3966093
Commit-Queue: Hugo Benichi <hugobenichi@google.com>
Tested-by: Hugo Benichi <hugobenichi@google.com>
Reviewed-by: Jintao Lin <jintaolin@chromium.org>