faced: Introduce FakeCameraClient implementation for testing
While we have a `FakeCameraService` class, the API it exposes is quite
low level, and difficult to test with. This CL adds a fake for the
higher-level `CameraClient` API.

In particular, this CL makes `CameraClient` an abstract interface with
two concrete implementations:

* `CrosCameraClient`, a real implementation that takes a CameraService

* `FakeCameraClient`, a fake implementation to simplify testing.

This CL does not actually use the FakeCameraClient (other than some
simple tests to exercise the fake), but upcoming CLs will.

BUG=b:237233809
TEST=`cros_workon_make --board=$BOARD --test chromeos-base/faced`

Cq-Depend: chrome-internal:5042862
Change-Id: Ibc1a2df348c4c1d43a09f8607370dd57931e3c72
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3963892
Reviewed-by: Ben Kersten <bkersten@google.com>
Reviewed-by: Victor Wong <vwon@google.com>
Commit-Queue: David Greenaway <dgreenaway@google.com>
Tested-by: David Greenaway <dgreenaway@google.com>