biod: Enforce finger up between auth attempts
If we allow starting a new auth session while we're still waiting for
finger up, we might consume a finger press for multiple matches in a
short period of time, using up all the allowed attempts. Therefore, we
have to enforce finger up between 2 auth attempts. When receiving a
StartAuthSession while state is WaitForFingerUp, transition the state
to AuthWaitForFingerUp, so that we process the AuthSession after finger
up event is detected.

BUG=b:251089506
TEST=biod unittests

Change-Id: I1837d39181ad003fbc29be6907ee5e43466c7aa8
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3966090
Tested-by: Howard Yang <hcyang@google.com>
Commit-Queue: Howard Yang <hcyang@google.com>
Reviewed-by: Craig Hesling <hesling@chromium.org>