cryptohome: Capture KeyData in MigrateKey
Currently, if you call MigrateKey it will overwrite the UserSession
KeyData with a blank message. This is probably technically incorrect by
itself, but it got much worse with the multiple-verifier support being
added because we depend on that key data to tell us which verifier to
update. This issue doesn't exist with the other (auth session and auth
factor) based APIs, which always construct and update proper KeyData
from their requests.

The quick fix for this is to also capture KeyData from the request,
similar to what the other older Key APIs (e.g. AddKey) do.

This also changes the CLI to populate the KeyData label from the
--key_label parameter, something which the various tests do actually set
when using migrate_key_ex, but which is actually currently just ignored
by the CLI tool.

BUG=b:254401604,b:254361667
TEST=FEATURES=test emerge-$BOARD cryptohome
TEST=tast run <dut> hwsec.CryptohomeKeysAndMountValidity

Change-Id: I365384d35ad0a69d49d6eedd2f742e4a7cda33b3
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3966930
Tested-by: John Admanski <jadmanski@chromium.org>
Reviewed-by: Betul Soysal <betuls@google.com>
Commit-Queue: John Admanski <jadmanski@chromium.org>
Reviewed-by: Maksim Ivanov <emaxx@chromium.org>