power: policy: wifi: Support multiple sensors like cellular
Use the same logic as cellular to set wifi power high or low
when several source of proximity exits (tablet mode and proximity
sensors):
When we are in tablet mode OR one proximity sensor indicates a user is
near we set wifi power to low.

BUG=b:251743138
TEST=Unit tests: `FEATURES=test emerge-brya power_manager`
Check on Crota360 we stay in low power mode when in
tablet mode, sitting on a table (off body kicks in).
Check on Shivana (Rammus, only tablet switch), we go low in power mode
on tablet mode.

Change-Id: Ia26b179a6c6407501a0f498f87d31e2c7a7ecaa7
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3965521
Tested-by: Scott Chao <scott_chao@wistron.corp-partner.google.com>
Reviewed-by: Matthew Wang <matthewmwang@chromium.org>
Commit-Queue: YH Lin <yueherngl@chromium.org>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>
Reviewed-by: Nicolas Norvez <norvez@chromium.org>