cryptohome: Add an async CredentialVerifier API
Due to the use of the verifier in (synchronous) legacy codepaths, we
can't just change the Verify API to be an async one. So instead this
compromises, by allowing verifiers to be implemented in either an async
or sync faction, depending on what's possible.

For verifiers that require async operations, like smart card and
fingerprint, they'll implement the async interface and will only be
usable via the newer AuthFactor APIs. For more basic verifiers
(password) that can operate synchronously, they'll implement the sync
interface and be usable via both old and new APIs.

This does also change the semantics of Authenticate in the case where
you have a verifier and it fails the auth. The existing code will then
fall back to trying to load an AuthBlock to authenticate, and using the
result of that as a final result. With this change, the result of the
verifier's Verify will be the final result, pass or fail.

BUG=b:247812443
TEST=FEATURES=test emerge-$BOARD cryptohome

Change-Id: I12ab959bbef3fab535534d2056be1ec16a833a43
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3965522
Commit-Queue: John Admanski <jadmanski@chromium.org>
Reviewed-by: Maksim Ivanov <emaxx@chromium.org>
Tested-by: John Admanski <jadmanski@chromium.org>