hiberman: Don't hibernate if system is updating
If the update engine is not idle, then refuse to hibernate the system.
It would be bad to hibernate if the "other" A/B slot were marked to boot
on the next reboot. If that happened, we'd still potentially resume from
hibernate, but would resume back into the world where the "old" rootfs
and kernel were running. This could severely confuse the system.
Additionally, firmware updates may also get applied during an update
reboot, changing the memory map passed to the kernel and voiding the
resume.

Add in an update_engine client call that uses the GetStatusAdvanced()
function to see if the update engine is busy. Only allow resumes if it's
idle.

BUG=b:237090293
TEST=Hibernate and resume, manually try failure case.

Cq-Depend: chromium:3969483
Change-Id: I227ad8fb2f0e2e616b8dedc037b032fe9c886f79
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3969699
Reviewed-by: David Jimenez <dvjimenez@chromium.org>
Commit-Queue: David Jimenez <dvjimenez@chromium.org>
Reviewed-by: Alexis Savery <asavery@chromium.org>
Tested-by: David Jimenez <dvjimenez@chromium.org>