secagentd: Add a process cache
The process cache tracks running as well as possibly recently exited
processes. It uses a LRU cache as its base and is currently configured
to store up to 256 processes. Uses the reporting::Process proto as the
cached object for easy CopyFrom() into various event types.

The cache is tightly coupled with ProcessPlugin as it relies on the
plugin to feed it with process_start events. However, the cache is also
able to trawl procfs to fill in missing processes on a best-effort
basis.

This cl just adds the class and unit tests. It will be integrated with
actual daemon code in a followup cl. Helper methods which appear to be
copy-pasted will also be removed from their original locations in a
followup cl.

BUG=b:241578161
TEST=Unit tests pass. Also tested on amd64-generic with WIP followup cl.

Change-Id: Ifbdc81e6a9f2800bd74b16fc6ce7f1fc951279e5
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3971423
Reviewed-by: Ryan Borzello <rborzello@google.com>
Reviewed-by: Jason Ling <jasonling@chromium.org>
Commit-Queue: Aashay Shringarpure <aashay@google.com>
Tested-by: Aashay Shringarpure <aashay@google.com>