patchpanel: RA server implementation, part 2
This patch adds support in ndproxy so that neighbor IP on a certain
downlink can be detected and trigger a callback without enabling RS/RA
proxy on the downlink. This functionality is used on the downlinks of
which RAs are provided by RA server.

BUG=b:228585272
TEST=`FEATURES=test emerge-$BOARD patchpanel`

Change-Id: I39951a58c6347a7dfb1d69cb086578b95804ed9d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3971819
Tested-by: Taoyu Li <taoyl@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Commit-Queue: Taoyu Li <taoyl@chromium.org>