oobe_config: Keep rollback data during reboot
In the past, rollback data was encrypted with owner key of the TPM.
It could be decrypted multiple times and would be sent to OOBE multiple
times.

Now we use pstore to store a secret, which only survives one reboot.

Before this CL, rollback data is decrypted, saved to encrypted stateful,
and then deleted once sent to OOBE. The information that rollback
happened would disappear upon reboot.

Stop deleting the decrypted data after every request from Chrome.
It will be deleted once the device is owned. Don't decrypt data on every
restore, but only if it hasn't been decrypted yet.

Why this is safe:
The only sensitive content in rollback data are network credentials.
Those are sent and configured on the first request from Chrome, and stay
configured until the TPM is cleared. The decrypted rollback data file is
also removed when the TPM is cleared, so we are not extending the window
when this data is available.

BUG=b:254803392
TEST=1) FEATURES=test emerge-hatch oobe_config
     2) tast run <dut> policy.EnterpriseRollbackInPlace
     3) Try it out manually:
       - touch /mnt/stateful_partition/.save_rollback_data
       - echo "fast safe keepimg rollback"
           > /mnt/stateful_partition/factory_install_reset
       - reboot
       - disconnect ethernet so device gets stuck in oobe
       - check logs, rollback data is decrypted and sent
       - reboot
       - check logs, rollback data is not decrypted, data from last time
	 is sent
       - crossystem clear_tpm_owner_request=1
       - reboot
       - check the device stays at welcome screen
       - check logs, encrypted rollback data was still around, device
	 tried to decrypt it but failed

Change-Id: Iab9b1c5988b3031d4f627b2721e6b4be40402a5c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3971101
Reviewed-by: Cristina Guerrero <crisguerrero@chromium.org>
Tested-by: Miriam Polzer <mpolzer@google.com>
Reviewed-by: Roland Bock <rbock@google.com>
Commit-Queue: Miriam Polzer <mpolzer@google.com>